# Conflicts:
#	comu/com.energyict.mdc.engine/src/main/java/com/energyict/mdc/engine/impl/core/JobExecution.java

+++ comu/com.energyict.mdc.device.data/src/main/java/com/energyict/mdc/device/data/impl/tasks/PriorityComTaskExecutionImpl.java
@@ -21,6 +21,7 @@ import com.energyict.mdc.common.tasks.ServerComTaskExecution;
 import com.energyict.mdc.common.tasks.TaskStatus;
 import com.energyict.mdc.common.tasks.history.ComTaskExecutionSession;
 
+import javax.xml.bind.annotation.XmlElement;
 import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
@@ -62,6 +63,11 @@ public class PriorityComTaskExecutionImpl implements PriorityComTaskExecution {
     @Override
     public Device getDevice() {
         return comTaskExecution.getDevice();
+    }
+
+    @Override
+    public void setDevice(Device device) {
+        comTaskExecution.setDevice(device);
     }
 
     @Override
@@ -277,6 +283,11 @@ public class PriorityComTaskExecutionImpl implements PriorityComTaskExecution {
     }
 
     @Override
+    public void setComTask(ComTask comTask) {
+        comTaskExecution.setComTask(comTask);
+    }
+
+    @Override
     public Optional<ComSchedule> getComSchedule() {
         return Optional.empty();
     }
@@ -299,7 +310,8 @@ public class PriorityComTaskExecutionImpl implements PriorityComTaskExecution {
     @Override
     public void executionFailed() {
         ((ComTaskExecutionImpl) comTaskExecution).executionFailed();
-        LOGGER.info("[high-prio] device: " + comTaskExecution.getDevice().getName() + "; comTaskExecution status: " + comTaskExecution.getStatus());        if (TaskStatus.Failed.equals(comTaskExecution.getStatus())) {
+        LOGGER.info("[high-prio] device: " + comTaskExecution.getDevice().getName() + "; comTaskExecution status: " + comTaskExecution.getStatus());
+        if (TaskStatus.Failed.equals(comTaskExecution.getStatus()) || TaskStatus.NeverCompleted.equals(comTaskExecution.getStatus())) {
             ((PriorityComTaskExecutionLinkImpl) comTaskExecutionLink).delete();
         } else {
             // execution attempt failed, will retry - should not delete yet the high-prio task
@@ -391,5 +403,16 @@ public class PriorityComTaskExecutionImpl implements PriorityComTaskExecution {
     @Override
     public void sessionCreated(ComTaskExecutionSession session) {
         comTaskExecution.sessionCreated(session);
+    }
+
+    @Override
+    @XmlElement(name = "type")
+    public String getXmlType() {
+        return this.getClass().getName();
+    }
+
+    @Override
+    public void setXmlType(String ignore) {
+        //Ignore, only used for JSON
     }
 }
